+++ backend/canisters/community/impl/src/jobs/mod.rs
@@ -1,3 +1,7 @@
+pub mod import_groups;
+
 use crate::RuntimeState;
 
-pub(crate) fn start(_runtime_state: &RuntimeState) {}
+pub(crate) fn start(state: &RuntimeState) {
+    import_groups::start_job_if_required(state);
+}