+++ backend/canisters/group_index/impl/src/guards.rs
@@ -7,3 +7,11 @@ pub fn caller_is_governance_principal() -> Result<(), String> {
         Err("Caller is not a governance principal".to_string())
     }
 }
+
+pub fn caller_is_community_canister() -> Result<(), String> {
+    if read_state(|state| state.is_caller_community_canister()) {
+        Ok(())
+    } else {
+        Err("Caller is not a community canister".to_string())
+    }
+}